+++ complex/library/changelog.md
@@ -1,5 +1,15 @@
 # WebMan WordPress Theme Framework Changelog
 
+## 1.0.11
+
+* **Fix**: Customizer custom option preview JS
+
+#### Files changed:
+
+	init.php
+	include/customize.php
+
+
 ## 1.0.10
 
 * **Update**: Removed custom color picker styling